+++ src/backend/demo/defaults.py
@@ -7,17 +7,12 @@ NB_OBJECTS = {
 }
 
 DEV_USERS = [
+    {"username": "impress", "email": "impress@impress.world", "language": "en-us"},
+    {"username": "user-e2e-webkit", "email": "user@webkit.e2e", "language": "en-us"},
+    {"username": "user-e2e-firefox", "email": "user@firefox.e2e", "language": "en-us"},
     {
-        "username": "impress",
-        "email": "impress@impress.world",
+        "username": "user-e2e-chromium",
+        "email": "user@chromium.e2e",
+        "language": "en-us",
     },
-    {
-        "username": "user-e2e-webkit",
-        "email": "user@webkit.e2e",
-    },
-    {
-        "username": "user-e2e-firefox",
-        "email": "user@firefox.e2e",
-    },
-    {"username": "user-e2e-chromium", "email": "user@chromium.e2e"},
 ]