+++ src/frontend/apps/impress/src/i18n/conf.ts
@@ -1,7 +1,0 @@
-export const LANGUAGES_ALLOWED: { [key: string]: string } = {
-  en: 'English',
-  fr: 'Français',
-  de: 'Deutsch',
-};
-export const LANGUAGE_COOKIE_NAME = 'docs_language';
-export const BASE_LANGUAGE = 'en';