+++ src/frontend/apps/impress/src/i18n/types.ts
@@ -1,2 +1,0 @@
-// See: https://github.com/numerique-gouv/impress/blob/ac58341984c99c10ebfac7f8bbe1e8756c48e4d4/src/backend/impress/settings.py#L156-L161
-export type ContentLanguage = 'en-us' | 'fr-fr';