+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+## Added
+
+## Changed
+
+- ✨(frontend) sync user and frontend language #401
+
+## Fixed
+
+- 🐛(backend) invitation e-mails in receivers language #401
+
 
 ## [1.8.2] - 2024-11-28